# src/frontend/src/lib/api/openapi.json
{
  "openapi" : "3.1.0",
  "components" : {
    "schemas" : {
...
      "ConflictDetailsDto" : {
        "type" : "object",
        "required" : [ "message", "blockingSequenceNumber", "blockingUserId", "blockingUserDisplayName", "blockingKind", "blockingCreatedAt", "currentDraftRevision" ],
        "properties" : {
          "message" : {
            "type" : "string"
          },
          "blockingSequenceNumber" : {
            "type" : "integer",
            "format" : "int64"
          },
          "blockingUserId" : {
            "$ref" : "#/components/schemas/UUID"
          },
          "blockingUserDisplayName" : {
            "type" : "string"
          },
          "blockingKind" : {
            "type" : "string"
          },
          "blockingCreatedAt" : {
            "$ref" : "#/components/schemas/Instant"
          },
          "currentDraftRevision" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
      "EstimationVersionSummaryDto" : {
        "type" : "object",
        "required" : [ "versionNumber", "isDraft", "totalEffort", "notes", "createdAt" ],
        "properties" : {
          "versionNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "isDraft" : {
            "type" : "boolean"
          },
          "totalEffort" : {
            "type" : [ "number", "null" ],
            "format" : "double"
          },
          "notes" : {
            "type" : [ "string", "null" ]
          },
          "createdAt" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/Instant"
            }, {
              "type" : "null"
            } ]
          }
        }
      },
      "Instant" : {
        "type" : "string",
        "format" : "date-time",
        "examples" : [ "2022-03-10T16:15:50Z" ]
      },
...
      "UUID" : {
        "type" : "string",
        "format" : "uuid",
        "pattern" : "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      },
...
    },
    "securitySchemes" : {
      "SecurityScheme" : {
        "description" : "Authentication",
        "type" : "openIdConnect",
        "openIdConnectUrl" : "https://login.microsoftonline.com/common/v2.0/.well-known/openid-configuration"
      }
    }
  },
...
  "paths" : {
...
    "/api/estimations/{estimationId}/versions" : {
      "get" : {
        "summary" : "List versions (live draft first, then submitted snapshots)",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The version summaries",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/EstimationVersionSummaryDto"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Estimation not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "VIEWER" ]
        } ]
      },
      "post" : {
        "summary" : "Create a draft version (cloned from the latest submitted, if any)",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "The created draft with calculated values",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EstimationVersionDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Estimation not found"
          },
          "409" : {
            "description" : "A draft already exists for this estimation"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/estimations/{estimationId}/versions/draft" : {
      "put" : {
        "summary" : "Replace the draft's collections wholesale and record the mutation",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DraftUpdateDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "The updated draft with recalculated values",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EstimationVersionDto"
                }
              }
            }
          },
          "404" : {
            "description" : "No draft found for this estimation"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          },
          "400" : {
            "description" : "Bad Request"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      },
      "get" : {
        "summary" : "Get the draft with on-the-fly calculated values",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The draft with calculated values",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EstimationVersionDto"
                }
              }
            }
          },
          "404" : {
            "description" : "No draft found for this estimation"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "VIEWER" ]
        } ]
      },
      "delete" : {
        "summary" : "Delete the draft version",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "The draft was deleted"
          },
          "404" : {
            "description" : "Estimation not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/estimations/{estimationId}/versions/draft/history" : {
      "get" : {
        "summary" : "The draft's mutation log (active and undone, ordered by sequence)",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The mutation log entries",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MutationLogEntryDto"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Estimation not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "VIEWER" ]
        } ]
      }
    },
    "/api/estimations/{estimationId}/versions/draft/redo" : {
      "post" : {
        "summary" : "Redo this user's last undone draft mutation",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The recalculated draft",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EstimationVersionDto"
                }
              }
            }
          },
          "404" : {
            "description" : "No draft found for this estimation"
          },
          "409" : {
            "description" : "A newer change blocks the redo",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConflictDetailsDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/estimations/{estimationId}/versions/draft/submit" : {
      "post" : {
        "summary" : "Snapshot the draft into an immutable submitted version",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The submitted version snapshot",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EstimationVersionDto"
                }
              }
            }
          },
          "404" : {
            "description" : "No draft found for this estimation"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/estimations/{estimationId}/versions/draft/undo" : {
      "post" : {
        "summary" : "Undo this user's last draft mutation",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The recalculated draft",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EstimationVersionDto"
                }
              }
            }
          },
          "404" : {
            "description" : "No draft found for this estimation"
          },
          "409" : {
            "description" : "A newer change blocks the undo",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConflictDetailsDto"
                }
              }
            }
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/estimations/{estimationId}/versions/import/merlin" : {
      "post" : {
        "summary" : "Import a Merlin project WBS as a new draft version (method-aware: tree for THREE_POINT_PERT, flattened bucketed leaves for BUCKET_SAMPLED_PERT)",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "required" : [ "file" ],
                "properties" : {
                  "file" : {
                    "type" : "string",
                    "format" : "binary"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "The created draft with calculated values",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EstimationVersionDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The upload is not a readable Merlin/SQLite file"
          },
          "404" : {
            "description" : "Estimation not found"
          },
          "409" : {
            "description" : "A draft already exists for this estimation"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "ESTIMATOR" ]
        } ]
      }
    },
    "/api/estimations/{estimationId}/versions/{versionA}/compare/{versionB}" : {
      "get" : {
        "summary" : "Diff two versions (use \"draft\" for the live draft)",
        "tags" : [ "Estimation versions" ],
        "parameters" : [ {
          "name" : "estimationId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/UUID"
          }
        }, {
          "name" : "versionA",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "versionB",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The version comparison",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/VersionComparisonDto"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced version was not found"
          },
          "401" : {
            "description" : "Not Authorized"
          },
          "403" : {
            "description" : "Not Allowed"
          }
        },
        "security" : [ {
          "SecurityScheme" : [ "VIEWER" ]
        } ]
      }
    },
...
  },
  "info" : {
    "title" : "Pythia API",
    "version" : "1.0.0-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}